Задача 4.4 (фикс 2, вынес FilterSettings в отдельный файл)
