+++ app/src/main/java/ru/practicum/android/diploma/ui/filter/FilterViewModel.kt
@@ -7,6 +7,7 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import ru.practicum.android.diploma.data.dto.FilterIndustryDto
+import ru.practicum.android.diploma.domain.model.FilterSettings
 import ru.practicum.android.diploma.domain.repository.FilterRepository
 import ru.practicum.android.diploma.util.Resource
 
@@ -121,16 +122,3 @@ class FilterViewModel(
     }
 }
 
-/**
- * Класс для хранения настроек фильтрации
- */
-data class FilterSettings(
-    val salary: Int? = null,
-    val onlyWithSalary: Boolean = false,
-    val industryId: Int? = null,
-    val industryName: String? = null,
-    val countryId: Int? = null,
-    val countryName: String? = null,
-    val regionId: Int? = null,
-    val regionName: String? = null
-)
